Add try/except checks in query

+++ application/server/views.py
@@ -4,7 +4,7 @@ import logging
 import mimetypes
 import os
 
-from pyramid.httpexceptions import HTTPFound
+from pyramid.httpexceptions import HTTPFound, HTTPNotFound, HTTPInternalServerError
 from pyramid.security import (
     remember,
     forget,
@@ -16,8 +16,6 @@ from server.models import (
     DBSession)
 from server.schemas import BannerSchema, LoginSchema
 from server.utils import crop_image
-
-from shutil import copyfile
 
 
 log = logging.getLogger(__name__)
@@ -50,9 +48,13 @@ class Views(object):
     @view_config(route_name='banners_view',
                  renderer='templates/banners_page.mako')
     def banners_view(self):
-        banners = DBSession.query(Banner).order_by(Banner.position, Banner.id)
+        try:
+            banners = DBSession.query(Banner).filter(Banner.visible == True).order_by(Banner.position, Banner.id).limit(15)
 
-        log.debug(200)
+        except Exception as e:
+            log.debug(e)
+            raise HTTPInternalServerError()
+
         return dict(banners=banners)
 
     @view_config(route_name='add_banner_view',
@@ -82,31 +84,34 @@ class Views(object):
                 visible=new_visible
             )
 
-            DBSession.add(new_banner)
+            try:
+                DBSession.add(new_banner)
 
-            banner = DBSession.query(Banner).filter(
-                Banner.title=new_title,
-                Banner.url=new_url,
-                Banner.visible=new_visible).order_by(Banner.id.desc()).first()
+                banner = DBSession.query(Banner).filter(
+                    Banner.title == new_title,
+                    Banner.url == new_url,
+                    Banner.visible == new_visible).order_by(Banner.id.desc()).first()
 
-            if appstruct.get("image") is None:
-                img_scr = ""
+                if appstruct.get("image") is None:
+                    img_scr = ""
 
-            else:
-                img_type = mimetypes.guess_extension(
-                    appstruct.get("image").get("mimetype"))
-                img_scr = f"static/banner_img/{banner.id}{img_type}"
+                else:
+                    img_type = mimetypes.guess_extension(
+                        appstruct.get("image").get("mimetype"))
+                    img_scr = f"static/banner_img/{banner.id}{img_type}"
 
-                crop_image(appstruct.get("image").get("fp"),
-                           f"server/{img_scr}")
+                    crop_image(appstruct.get("image").get("fp"),
+                            f"server/{img_scr}")
 
-            banner.image_path = img_scr
-            banner.position = banner.id
+                banner.image_path = img_scr
+                banner.position = banner.id
 
-            log.debug(201)
+                url = self.request.route_url('banners_view')
+                return HTTPFound(url)
 
-            url = self.request.route_url('banners_view')
-            return HTTPFound(url)
+            except Exception as e:
+                log.debug(e)
+                raise HTTPInternalServerError()
 
         log.debug(200)
         return dict(form=form)
@@ -115,14 +120,19 @@ class Views(object):
     def delete_banner_view(self):
         bid = int(self.request.matchdict['id'])
 
-        banner = DBSession.query(Banner).filter(Banner.id == bid).first()
+        try:
+            banner = DBSession.query(Banner).filter(Banner.id == bid).first()
 
-        if os.path.exists(f"server/{banner.image_path}" or ""):
-            os.remove(f"server/{banner.image_path}")
+            if banner is None:
+                raise HTTPNotFound
 
-        DBSession.delete(banner)
+            if os.path.exists(f"server/{banner.image_path}" or ""):
+                os.remove(f"server/{banner.image_path}")
 
-        log.debug(201)
+            DBSession.delete(banner)
+
+        except Exception as e:
+            log.debug(e)
 
         url = self.request.route_url('banners_view')
         return HTTPFound(url)
@@ -133,7 +143,15 @@ class Views(object):
     def update_banner_view(self):
         bid = int(self.request.matchdict['id'])
 
-        banner = DBSession.query(Banner).filter(Banner.id == bid).first()
+        try:
+            banner = DBSession.query(Banner).filter(Banner.id == bid).first()
+
+            if banner is None:
+                raise HTTPNotFound
+
+        except Exception as e:
+            log.debug(e)
+            raise HTTPInternalServerError
 
         form = self.banner_form.render({
             "title": banner.title,
@@ -144,14 +162,22 @@ class Views(object):
 
         image_path = f"server:{banner.image_path}"
 
-        form = form.replace('<label for="deformField2"\n         \
-            class=\"control-label "\n         id="req-deformField2"\n         \
-                > \\n    Image\n  </label>',
-                            f'<label for="deformField2"\n         \
-            class="control-label "\n         id="req-deformField2"\n         >\
-                \n    Image\n  </label>\n<br><img src=\"\
-                {self.request.static_url(image_path)}\" alt=\"\" \
-                    draggable="false" width=200 height=200/><br>\n')
+        try:
+            form = form.replace('<label for="deformField2"\n         ' +
+                'class=\"control-label "\n         id="req-deformField2"\n         ' +
+                    '> \\n    Image\n  </label>',
+                                f'<label for="deformField2"\n         \
+                class="control-label "\n         id="req-deformField2"\n         >\
+                    \n    Image\n  </label>\n<br><img src=\"\
+                    {self.request.static_url(image_path)}\" alt=\"\" \
+                        draggable="false" width=200 height=200/><br>\n')
+
+        except ValueError:
+            pass
+
+        except Exception as e:
+            log.debug(e)
+            raise HTTPInternalServerError
 
         if 'submit' in self.request.params:
             controls = self.request.POST.items()
@@ -159,10 +185,8 @@ class Views(object):
             try:
                 appstruct = self.banner_form.validate(controls)
 
-            except deform.ValidationFailure as e:
-
-                log.debug(400)
-                return dict(form=e.render())
+            except deform.ValidationFailure as ve:
+                return dict(form=ve.render())
 
             new_title = appstruct.get("title", "default")
             new_url = appstruct.get("url", "default")
@@ -183,13 +207,16 @@ class Views(object):
             else:
                 img_scr = banner.image_path
 
-            banner.title = new_title
-            banner.image_path = img_scr
-            banner.url = new_url
-            banner.visible = new_visible
-            banner.updated_at = datetime.datetime.utcnow()
+            try:
+                banner.title = new_title
+                banner.image_path = img_scr
+                banner.url = new_url
+                banner.visible = new_visible
+                banner.updated_at = datetime.datetime.utcnow()
 
-            log.debug(201)
+            except Exception as e:
+                log.debug(e)
+                raise HTTPInternalServerError
 
             url = self.request.route_url('banners_view')
             return HTTPFound(url)
@@ -207,8 +234,6 @@ class Views(object):
                 appstruct = self.login_form.validate(controls)
 
             except deform.ValidationFailure as e:
-
-                log.debug(400)
                 return dict(form=e.render())
 
             request = self.request
@@ -223,7 +248,6 @@ class Views(object):
             name = appstruct.get("name")
             headers = remember(request, name)
 
-            log.debug(201)
             return HTTPFound(location=came_from,
                              headers=headers)
 
@@ -236,7 +260,6 @@ class Views(object):
         headers = forget(request)
         url = request.route_url('banners_view')
 
-        log.debug(201)
         return HTTPFound(location=url,
                          headers=headers)
 
@@ -244,20 +267,39 @@ class Views(object):
                  permission='admin')
     def increase_banner_position_view(self):
         bid = int(self.request.matchdict['id'])
-        cursor_banner = DBSession.query(Banner).filter(Banner.id == bid).first()
 
-        banners = DBSession.query(Banner).order_by(Banner.position, Banner.id).all()
+        try:
+            cursor_banner = DBSession.query(Banner).filter(Banner.id == bid).first()
+
+            if cursor_banner is None:
+                raise HTTPNotFound
+
+        except Exception as e:
+            log.debug(e)
+            raise HTTPInternalServerError
+
+        try:
+            banners = DBSession.query(Banner).order_by(Banner.position, Banner.id).all()
+
+        except Exception as e:
+            log.debug(e)
+            raise HTTPInternalServerError
 
         bindex = banners.index(cursor_banner)
 
         if bindex != 0:
             cursor_position = banners[bindex].position
 
-            banners[bindex].position = banners[bindex - 1].position
-            banners[bindex].updated_at = datetime.datetime.utcnow()
+            try:
+                banners[bindex].position = banners[bindex - 1].position
+                banners[bindex].updated_at = datetime.datetime.utcnow()
 
-            banners[bindex - 1].position = cursor_position
-            banners[bindex - 1].updated_at = datetime.datetime.utcnow()
+                banners[bindex - 1].position = cursor_position
+                banners[bindex - 1].updated_at = datetime.datetime.utcnow()
+
+            except Exception as e:
+                log.debug(e)
+                raise HTTPInternalServerError
 
         log.debug(201)
         url = self.request.route_url('banners_view')
@@ -267,20 +309,39 @@ class Views(object):
                  permission='admin')
     def decrease_banner_position_view(self):
         bid = int(self.request.matchdict['id'])
-        cursor_banner = DBSession.query(Banner).filter(Banner.id == bid).first()
 
-        banners = DBSession.query(Banner).order_by(Banner.position, Banner.id).all()
+        try:
+            cursor_banner = DBSession.query(Banner).filter(Banner.id == bid).first()
+
+            if cursor_banner is None:
+                raise HTTPNotFound
+
+        except Exception as e:
+            log.debug(e)
+            raise HTTPInternalServerError
+
+        try:
+            banners = DBSession.query(Banner).order_by(Banner.position, Banner.id).all()
+
+        except Exception as e:
+            log.debug(e)
+            raise HTTPInternalServerError
 
         bindex = banners.index(cursor_banner)
 
         if bindex + 1 != len(banners):
             cursor_position = banners[bindex].position
 
-            banners[bindex].position = banners[bindex + 1].position
-            banners[bindex].updated_at = datetime.datetime.utcnow()
+            try:
+                banners[bindex].position = banners[bindex + 1].position
+                banners[bindex].updated_at = datetime.datetime.utcnow()
 
-            banners[bindex + 1].position = cursor_position
-            banners[bindex + 1].updated_at = datetime.datetime.utcnow()
+                banners[bindex + 1].position = cursor_position
+                banners[bindex + 1].updated_at = datetime.datetime.utcnow()
+
+            except Exception as e:
+                log.debug(e)
+                raise HTTPInternalServerError
 
         log.debug(201)
         url = self.request.route_url('banners_view')
